API

AFL.Result — Type

Result

Enum representing the return state of the running the target with an input.

AFL.Target — Type

Target

Represents a running AFL forkserver and a shared memory segment for the branch bitmap. Create one with init_target and close the handle with close.

AFL.CoverageMap — Type

CoverageMap

An array of bytes used to describe the coverage an input generated. Each byte represents a specific edge between in the code that can be covered, each bit represents a classification of counts this edge was taken.

AFL.InvCoverageMap — Type

InvCoverageMap

See CoverageMap. This is an inverted version, which can be faster in some cases. Often used to describe edges and counts not covered yet.

AFL.init_target — Function

init_target(target_path::String; qemu=false, memlimit=200)

Start a forkserver for the given target.

AFL.run_target — Function

run_target(target::Target, input::Vector{UInt8}; timeout=1)

Run the target with the specified input. Times out after timeout seconds. Return value is a Result enum.

AFL.classify_coverage — Function

classify_coverage(target::Target)

Convert the trace bits in the shared memory segment of the target to a classified coverage map.

AFL.classify_coverage! — Function

classify_coverage!(dst::CoverageMap, target::Target)

In place version of classify_coverage.

AFL.coverage_stats — Function

edges_covered, classes_covered = coverage_stats(map::CoverageMap)
edges_covered, classes_covered = coverage_stats(map::InvCoverageMap)

Calculate the ratio of edges and classes covered in the give coverage map. The maximum possible coverage (1024 * 64 edges covered) is represented as 1.0.

AFL.has_new_coverage — Function

has_new_coverage(dst::InvCoverageMap, src::CoverageMap)

Compares dst with src and returns true when src covers edges or classes not yet marked as covered in src, otherwise returns false.

AFL.mark_covered! — Function

mark_covered!(dst::InvCoverageMap, src::CoverageMap)

Marks the edges and classes covered by src as covered in dst.